requirements.txt
black==19.10b0
coverage==4.5.2
Django==2.2.3
django-cors-headers==3.7.0
djangorestframework==3.10.1
flake8==3.6.0
isort==4.3.4
pytz==2018.9
virtualenv==16.3.0
